requirements.txt
# Neural BUS Requirements
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Networking libraries (avoid LibreSSL warnings)
urllib3<2